Bug 1177798 - Only cache a dictionary of user data instead of full ORM objects.
This should reduce the amount spent in Python a bunch.
