+++ kuma/users/tests/test_helpers.py
@@ -3,12 +3,10 @@ from hashlib import md5
 
 from django.conf import settings
 
-from jinja2 import Markup
 from nose.tools import eq_, ok_
-from pyquery import PyQuery as pq
 
-from . import UserTestCase, user
-from ..helpers import gravatar_url, public_email, user_list
+from . import UserTestCase
+from ..helpers import gravatar_url, public_email
 
 
 class HelperTestCase(UserTestCase):
@@ -19,12 +17,12 @@ class HelperTestCase(UserTestCase):
 
     def test_profile_default_gravatar(self):
         d_param = urllib.urlencode({'d': settings.DEFAULT_AVATAR})
-        ok_(d_param in gravatar_url(self.u),
-            "Bad default avatar: %s" % gravatar_url(self.u))
+        ok_(d_param in gravatar_url(self.u.email),
+            "Bad default avatar: %s" % gravatar_url(self.u.email))
 
     def test_gravatar_url(self):
         self.u.email = 'test@test.com'
-        ok_(md5(self.u.email).hexdigest() in gravatar_url(self.u))
+        ok_(md5(self.u.email).hexdigest() in gravatar_url(self.u.email))
 
     def test_public_email(self):
         eq_('<span class="email">'
@@ -33,14 +31,3 @@ class HelperTestCase(UserTestCase):
         eq_('<span class="email">'
             '&#110;&#111;&#116;&#46;&#97;&#110;&#46;&#101;&#109;&#97;&#105;'
             '&#108;</span>', public_email('not.an.email'))
-
-    def test_user_list(self):
-        user(pk=400000, username='testuser3', save=True)
-        users = self.user_model.objects.all()
-        list = user_list(users)
-        assert isinstance(list, Markup)
-        fragment = pq(list)
-        eq_(5, len(fragment('a')))
-        a = fragment('a')[4]
-        assert a.attrib['href'].endswith('testuser3')
-        eq_('testuser3', a.text)
