+++ kuma/wiki/templates/wiki/document.html
@@ -1,7 +1,7 @@
 {% extends "wiki/base.html" %}
 {% block title %}{{ page_title(document.title + seo_parent_title) }}{% endblock %}
 
-{% from "wiki/includes/document_macros.html" import build_document_crumbs, get_document_quick_links, get_document_buttons, get_document_subnav, get_approvals_html, document_watch, share_links with context %}
+{% from "wiki/includes/document_macros.html" import build_document_crumbs, get_document_quick_links, get_document_buttons, get_document_subnav, get_approvals_html, document_watch, share_links, contributor_links with context %}
 
 {% if fallback_reason == 'no_translation' %}
   {% set help_link = url('wiki.translate', document_path=document.slug, locale=document.locale)|urlparams(tolocale=request.locale) %}
@@ -180,8 +180,9 @@
         {% endif %}
 
         {% if waffle.flag('top_contributors') %}
+            {% set contributors = document.get_contributors() %}
+            {% set contrib_count = contributors|length %}
             {% set contrib_limit = 13 %}
-            {% set contrib_count = document.contributors.count() %}
 
         <div class="contributor-avatars" data-all-text="{{ _('Show all') }}&hellip;<span class='hidden'>{{ _('contributors') }}</span>" {% if contrib_count > contrib_limit %}data-has-hidden="1"{% endif %}>
             <span class="quickstat">
@@ -192,10 +193,10 @@
             {% endtrans %}
             </span>
             <ul>
-            {% for contributor in document.contributors %}
+            {% for contributor in contributors %}
                 <li class="{% if loop.index > contrib_limit %}hidden{% else %}shown{% endif %}">
-                <a href="{{ contributor.get_absolute_url() }}" title="{% trans username=contributor.username %}View profile: {{ username }}{% endtrans %}">
-                <noscript data-class="avatar" data-alt="{{ contributor.username }}" data-src="{{ gravatar_url(contributor, size=34) }}">{{ contributor.username }}</noscript></a>
+                <a href="{{ url('users.profile', username=contributor.username) }}" title="{% trans username=contributor.username %}View profile: {{ username }}{% endtrans %}">
+                <noscript data-class="avatar" data-alt="{{ contributor.username }}" data-src="{{ contributor.gravatar_34 }}">{{ contributor.username }}</noscript></a>
                 </li>
             {% endfor %}
             </ul>
@@ -355,7 +356,7 @@
                   <h2 class="offscreen">{{ _('Document Tags and Contributors') }}</h2>
                   {% set tags = document.tags.all().order_by('name') %}
                   {% include "wiki/includes/document_tag.html" %}
-                  {% trans contributors_list=user_list(contributors) %}
+                  {% trans contributors_list=contributor_links(contributors) %}
                   <div class="contributors-sub">
                     <i aria-hidden="true" class="icon-group"></i><strong>Contributors to this page:</strong> {{ contributors_list }}
                   </div>
